+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 
+- Remove nested foreign keys from open api schema for methods POST PUT and PATCH
 - Rename unittests from `tests_` to `test_`
 - sets extra fields: Check that the field is in fields before checking the type